+++ SC.NewbLibrary/src/Pages/FetchData.razor
@@ -1,6 +1,7 @@
 ﻿@page "/fetchdata"
 
-@using SC.NewbLibrary.Data
+@using SC.NewbLibrary.Model.Data
+@using SC.NewbLibrary.Service
 @inject WeatherForecastService ForecastService
 
 <h1>Weather forecast</h1>
