+++ SC.NewbLibrary/src/Pages/Terms.razor
@@ -1,6 +1,7 @@
 ﻿@page "/terms"
 
-@using SC.NewbLibrary.Data
+@using SC.NewbLibrary.Model.Data
+@using SC.NewbLibrary.Service
 @inject NewbDataService NewbService
 
 <h1>Terms</h1>
